backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
pydantic
python-dateutil
bcrypt
google-auth
requests
